Move Future to kas::future module
